chore: fixes the HashLeaf description (#92)
## Overview

While reviewing the nmt codebase (as part of this
[EPIC](https://github.com/celestiaorg/celestia-app/issues/1296)) I
realized the description of the `HashLeaf` function does not match its
implementation (w.r.t. referred variable names) which causes confusion
when comprehending the code. This PR aims to correct and revise that
description.

Also, the illustration of an nmt provided in the [Readme
file](https://github.com/celestiaorg/nmt/blob/master/README.md) seems
not aligned with the leaf hash calculation, where in the `node_0.0` the
`d_0.0=H(data_0)` while it should be `d_0.0=H(0x00, nid_0, data_0)`. I
was not able to edit the image as its source code is not available.

Co-authored-by: sanaz <sanaztaheri@ieee.org>