chore: NMT specifications (#101)
## Overview
This pull request aligns with the
https://github.com/celestiaorg/celestia-app/issues/1296 EPIC and
introduces an NMT specification that includes a description of NMT as a
data structure, an explanation of the NMT library, and a sample code. In
terms of description of the NMT library, the focus was on the portions
that are used in the NMT wrapper to establish the foundation for the NMT
wrapper specification. The NMT methods are illustrated through a running
code example that was originally included in the Readme file of the NMT
repository (with a minor change in namespace ID of one of the leaves in
order to be able to provide an example of absence proof). This example
has been expanded to provide a more detailed description of each method
call, an interpretation of the parameters, and its relationship to the
NMT high-level logic. The The corresponding NMT for the code example is
also visualized to facilitate a better understanding and examination of
the expected behavior of the library.