feat: verify the range of the namespace IDs of the children in the HashNode (#102)
## Overview
This PR closes #95. 
Inline with https://github.com/celestiaorg/celestia-app/issues/1296.

The implementation of this pull request does not exactly follow the
original suggestion in the LazyLedge paper. Specifically, it does not
treat `leftMaxNs = rightMinNs` as a violation of leaf ordering. The
reason for this is that a namespace may span across multiple subtrees.
cc: @liamsi  

Quoting from the article:
> An adversarial consensus node may attempt to produce a block that
contains a Merkle tree with children that are not ordered correctly. To
prevent this, we can set a condition in nsHash such that there is no
valid hash when **leftMaxNs ≥ rightMinNs**, and thus there would be no
valid Merkle root for incorrectly ordered children

## Checklist

- [x] New and updated code has appropriate documentation
- [x] New and updated code has new and/or updated testing
- [x] Required CI checks are passing
- [x] Visual proof for any user facing features like CLI or
documentation updates
- [x] Linked issues closed with keywords