test: ExampleNamespacedMerkleTree (#203)
Previously the function `ExampleNamespacedMerkleTree` existed in
nmt_test.go. AFAIK Go only runs unit tests for functions with the name
`Test*` so I wouldn't expect `ExampleNamespacedMerkleTree` to run by the
Go test runner. But it is run and fails
[here](https://github.com/celestiaorg/nmt/actions/runs/5092193098/jobs/9153228914?pr=202#step:7:551).

This PR converts the example into a test. Defer to @staheri14 if it
should be deleted instead.