feat(proof.go): adding JSON marshal/unmarshal (#214)
Adds json marshaling/unmarshaling to `nmt.Proof`.

As found elsewhere in the go stdlib, `UnmarshalJSON` must be defined on
the pointer receiver although `MarshalJSON` is a value receiver.

This is needed because the fields are private. Does this need to be the
case? If we make them public, JSON marshaling will work by default.