feat!: adds input range check to optimize VerifyLeafHashes and VerifyInclusion methods (#253)
This PR optimizes the `VerifyLeafHashes` and `VerifyInclusion` methods
by ensuring that the size of the provided leaves or leaf hashes matches
the proof range and, if not, making an early return in order to prevent
unnecessary hashing operations that would otherwise occur.
It introduces a breaking change by altering the behaviour of
`VerifyLeafHashes` and `VerifyInclusion`. Previously, verification would
succeed even if the provided leaves or leaf hashes exceeded the proof
range, as long as the proof was valid and the extra leaves were appended
at the end. In the new implementation, the verification will return
false or an error in such cases.